feat(client): replaced obsolete proxy factory
